Moved _ensure_unique().
